exception handler 추가하여 오류 메세지 반환하게끔 수정, delete 시 비밀번호 요구하게끔 변경

+++ spring/src/main/java/com/jungle/week13/service/BoardService.java
@@ -54,7 +54,19 @@ public class BoardService {
         return post.getPostId();
     }
 
-    public Long deletePost(Long id) {
+    public Long deletePost(Long id, PostRequestDto requestDto) {
+        Post post = getPost(id);
+        String oldPostPwd = post.getPostPwd();
+        String postPwd = requestDto.getPostPwd();
+
+        if (StringUtils.isEmpty(postPwd)) {
+            throw new IllegalArgumentException("비밀번호 값이 입력되지 않았습니다.");
+        }
+
+        if (!StringUtils.equals(oldPostPwd, postPwd)) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
         boardRepository.deleteById(id);
         return id;
     }
